+++ CHANGELOG.md
@@ -5,6 +5,9 @@ FEATURES:
 FEATURES:
 * Adds `tfe_query_run` action, allowing users to invoke remote Terraform queries on HCP Terraform and Terraform Enterprise, by @sebasslash [#1982](https://github.com/hashicorp/terraform-provider-tfe/pull/1982)
 
+BUG FIXES:
+* `r/tfe_team_notification_configuration`: Added the `token_wo_version` write-only attribute to allow triggering `token_wo` updates, by @uk1288 ([#1995](https://github.com/hashicorp/terraform-provider-tfe/pull/1995))
+
 ## v0.74.1
 
 BUG FIXES: